Move stack struct to util/containers pkg
