+++ util/containers/stack.go
@@ -1,7 +1,7 @@
 // Copyright 2024, Offchain Labs, Inc.
 // For license information, see https://github.com/OffchainLabs/nitro/blob/master/LICENSE
 
-package stack
+package containers
 
 import (
 	"fmt"
